requirements.txt
yfinance==0.1.74
eodhd==1.0.7
python-dotenv==0.21.0
